Separate control content graph as component to swap for SPloM
