+++ src/components/ControlContent.vue
@@ -1,0 +1,117 @@
+<script>
+import * as d3 from "d3";
+export default {
+    name: "ControlContent",
+    components: {},
+    props: {},
+    data() {
+        return {
+            shouldShowControls: true,
+        };
+    },
+    mounted() {
+        const width = 800;
+        const height = 100;
+        const data = [
+            { date: "24-Apr-07", amount: 93.24 },
+            { date: "25-Apr-07", amount: 95.35 },
+            { date: "26-Apr-07", amount: 98.84 },
+            { date: "27-Apr-07", amount: 99.92 },
+            { date: "30-Apr-07", amount: 99.8 },
+            { date: "1-May-07", amount: 99.47 },
+            { date: "2-May-07", amount: 100.39 },
+            { date: "3-May-07", amount: 100.4 },
+            { date: "4-May-07", amount: 100.81 },
+            { date: "7-May-07", amount: 103.92 },
+            { date: "8-May-07", amount: 105.06 },
+            { date: "9-May-07", amount: 106.88 },
+            { date: "10-May-07", amount: 107.34 },
+        ];
+
+        const svg = d3.select("#selectionPlot")
+            .attr("width", width) 
+            .attr("height", height + 25) //add 25 to accomidate the x axis labels
+            .attr("stroke", "pink");// set current color to pink
+        const g = svg.append("g");
+
+        //2. Parse the dates
+        const parseTime = d3.timeParse("%d-%b-%y");
+
+        //3. Creating the Chart Axes
+        const x = d3
+            .scaleTime()
+            .domain(
+                d3.extent(data, function (d) {
+                    return parseTime(d.date);
+                })
+            )
+            .rangeRound([0, width]);
+
+        const y = d3
+            .scaleLinear()
+            .domain(
+                d3.extent(data, function (d) {
+                    return d.amount;
+                })
+            )
+            .rangeRound([height, 0]);
+
+        //4. Creating a Line
+        const line = d3
+            .line()
+            .x(function (d) {
+                return x(parseTime(d.date));
+            })
+            .y(function (d) {
+                return y(d.amount);
+            });
+
+        //5. Appending the Axes to the Chart
+        g.append("g")
+            .call(d3.axisBottom(x))
+            .attr("transform", "translate(0," + (height) + ")")
+            .append("text")
+            .attr("fill", "pink")
+            .attr("font-size", 10)
+            .attr("x", "11em")
+            .attr("y", -6)
+            .attr("text-anchor", "start")
+            .text("Time (t)")
+
+        g.append("g")
+            .call(d3.axisLeft(y))
+            .append("text")
+            .attr("fill", "pink")
+            .attr("transform", "rotate(-90)")
+            .attr("y", 6)
+            .attr("dy", "0.71em")
+            .attr("text-anchor", "end")
+            .text("Price ($)");
+
+        //6. Appending a path to the Chart
+        g.append("path")
+            .datum(data)
+            .attr("fill", "none")
+            .attr("stroke", "pink")
+            .attr("stroke-width", 2.5)
+            .attr("d", line);
+        g.selectAll(".domain,.tick,line")
+            .style("stroke", "pink");
+    },
+};
+</script>
+<template>
+  <svg id="selectionPlot"></svg>
+</template>
+<style scoped>
+  .font-semibold{
+    font-weight: 600;
+  }
+  .text-xl{
+    font-size: 1.25rem; /* 20px */
+    line-height: 1.75rem; /* 28px */
+  }
+  .no-wrap{
+    white-space: nowrap;
+  }
+</style>
